fix: getIsPremiumContent 수정
